refactor(discovery): use the canonical CollapsibleList, not a hand-rolled toggle
Replaces the AttributedEstateGrid client component added earlier in this branch.

docs/platform-usability-standards.md is explicit: "Do not hand-roll a new
expand/collapse dialect inside a feature", and names `CollapsibleList` as the
canonical construct for exactly this need — "preview the first rows of one long
list". I had written a sixth divergent implementation of the thing that
primitive exists to prevent; its own header notes it was created because ~5
surfaces had each hand-rolled useState + slice(0, N) + a bespoke toggle.

Three concrete wins beyond conformance:

  - `CollapsibleList` emits `data-dpf-disclosure-trigger`, which the UX budget
    measurement recognises as a sanctioned disclosure region. The hand-rolled
    version would never have been credited for deferring anything — it would
    have shrunk the page while still measuring as if everything were visible.
  - It takes already-rendered children, so DiscoveryOperationsPage STAYS A
    SERVER COMPONENT. The hand-rolled version forced a new client boundary
    purely to hold toggle state.
  - It carries the WAI-ARIA disclosure contract (aria-expanded/aria-controls,
    trigger always at the foot of the list) that my version only partly had.

Net: 37 added source lines instead of 134, four files instead of seven, no new
component, no new client boundary, and the behaviour is covered by
CollapsibleList's own tests rather than a parallel suite.

Kept from the previous approach: the 12-card preview, the shared guidance
sentence stated once instead of repeated on all 381 cards, and the migration of
arbitrary text-[9px]/[10px] sizes to the generated text-dpf-caption scale.

Verification: 114 tests pass across components/inventory, the discovery route,
and report-kit. apps/web typechecks clean. Module-size, style/token drift,
design-grounding and prose-lint all green; the stale prose-lint baseline entry
for the deleted component is removed.
