+++ apps/web/components/inventory/DiscoveryOperationsPage.tsx
@@ -1,6 +1,8 @@
 import Link from "next/link";
 import { INVENTORY_ENTITY_CANONICAL_WHERE, prisma } from "@dpf/db";
 
+import { PORTFOLIO_COLOURS } from "@/lib/portfolio";
+
 import {
   countStaleEntitiesSince,
   getInventoryEntitiesGroupedBySubnet,
@@ -18,7 +20,24 @@ import { SavedConnectionsPanel } from "@/components/inventory/SavedConnectionsPa
 import { SubnetGroupedInventoryPanel } from "@/components/inventory/SubnetGroupedInventoryPanel";
 import { TopologyGraph } from "@/components/inventory/TopologyGraph";
 import { CustomerTopologyScopeBar } from "@/components/inventory/CustomerTopologyScopeBar";
-import { AttributedEstateGrid } from "@/components/inventory/AttributedEstateGrid";
+import { CollapsibleList } from "@/components/ui/report-kit/CollapsibleList";
+
+const STATUS_COLOURS: Record<string, string> = {
+  active: "var(--dpf-success)",
+  draft: "var(--dpf-warning)",
+  inactive: "var(--dpf-muted)",
+};
+
+/**
+ * Product cards visible before the estate grid defers the rest.
+ *
+ * One card per DigitalProduct carrying discovery evidence: 381 cards / 20,572px
+ * on the live install — two thirds of a page that stood 34 screens tall. This
+ * surface resolves evidence quality and then hands off (its own header says to
+ * "manage the owned estate from portfolio and product pages"), so the full
+ * roster belongs behind the disclosure rather than in the arrival viewport.
+ */
+const ESTATE_PREVIEW_COUNT = 12;
 
 type DiscoveryOperationsPageProps = {
   isLegacyAlias?: boolean;
@@ -171,15 +190,62 @@ export async function DiscoveryOperationsPage({
           </Link>
         </div>
 
-        <AttributedEstateGrid
-          products={products.map((product) => ({
-            id: product.id,
-            name: product.name,
-            lifecycleStatus: product.lifecycleStatus,
-            portfolio: product.portfolio,
-            taxonomyNodeId: product.taxonomyNode?.nodeId ?? null,
-          }))}
-        />
+        <p className="text-dpf-caption text-[var(--dpf-muted)]">
+          Open a product to review dependencies, supporting items, and posture in context.
+        </p>
+
+        <CollapsibleList
+          as="div"
+          previewCount={ESTATE_PREVIEW_COUNT}
+          listClassName="grid grid-cols-1 gap-3 sm:grid-cols-2"
+          moreLabel={(hidden) => `Show ${hidden} more products`}
+          fewerLabel="Show fewer products"
+        >
+          {products.map((product) => {
+            const colour = product.portfolio
+              ? (PORTFOLIO_COLOURS[product.portfolio.slug] ?? "var(--dpf-accent)")
+              : "var(--dpf-border)";
+            const statusColour = STATUS_COLOURS[product.lifecycleStatus] ?? "var(--dpf-muted)";
+            const taxonomyPath = product.taxonomyNode
+              ? product.taxonomyNode.nodeId.replace(/\//g, " / ")
+              : null;
+
+            return (
+              <Link
+                key={product.id}
+                href={`/portfolio/product/${product.id}/inventory`}
+                className="block rounded-lg border border-[var(--dpf-border)] bg-[var(--dpf-surface-1)] p-4 transition-colors hover:border-[var(--dpf-accent)]"
+                style={{ borderLeft: `4px solid ${colour}` }}
+              >
+                <div className="flex items-start justify-between gap-2">
+                  <div>
+                    <p className="text-sm font-semibold text-[var(--dpf-text)]">{product.name}</p>
+                    {product.portfolio && (
+                      <p className="mt-1 text-dpf-caption font-medium" style={{ color: colour }}>
+                        {product.portfolio.name}
+                      </p>
+                    )}
+                  </div>
+                  <span
+                    className="rounded-full px-1.5 py-0.5 text-dpf-caption"
+                    style={{ backgroundColor: `${statusColour}20`, color: statusColour }}
+                  >
+                    {product.lifecycleStatus}
+                  </span>
+                </div>
+                {taxonomyPath && (
+                  <p className="mt-2 text-dpf-caption font-mono text-[var(--dpf-muted)]">{taxonomyPath}</p>
+                )}
+              </Link>
+            );
+          })}
+        </CollapsibleList>
+
+        {products.length === 0 && (
+          <p className="text-sm text-[var(--dpf-muted)]">
+            No products are linked to discovered estate evidence yet.
+          </p>
+        )}
       </div>
     </div>
   );
